+++ apps/web/src/app/api/auth/register/route.ts
@@ -7,6 +7,8 @@ import { Prisma } from "@prisma/client";
 import { SendVerification } from "@/lib/mail";
 
 const normalizeEmail = (email: string) => email.trim().toLowerCase();
+const IGNORE_EMAIL_VERIFY_string = process.env.IGNORE_EMAIL_VERIFY ?? "";
+const IGNORE_EMAIL_VERIFY = (IGNORE_EMAIL_VERIFY_string === "1" || IGNORE_EMAIL_VERIFY_string.toUpperCase() === "TRUE")
 
 export async function POST(request: Request) {
   let body: { email?: string; password?: string; displayName?: string };
@@ -29,9 +31,8 @@ export async function POST(request: Request) {
   if (password.length < 8)
     return NextResponse.json({ error: "Password must be at least 8 characters." }, { status: 400 });
 
-  const IGNORE_EMAIL_VERIFY = process.env.IGNORE_EMAIL_VERIFY ?? "";
 
-  if (IGNORE_EMAIL_VERIFY === "1" || IGNORE_EMAIL_VERIFY.toUpperCase() === "TRUE") {
+  if (IGNORE_EMAIL_VERIFY) {
     try {
       const passwordHash = await passwordHashing(password);
 
@@ -64,17 +65,17 @@ export async function POST(request: Request) {
     } catch (e) {
       if (e instanceof Prisma.PrismaClientKnownRequestError && e.code === "P2002")
         return NextResponse.json({ error: "Email already in use." }, { status: 409 });
-      return NextResponse.json({ error: "Internal server error." }, { status: 500 });
+      else
+        return NextResponse.json({ error: "Internal server error." }, { status: 500 });
     }
   }
 
   try {
-    if (
-      await SendVerification({
-        stringVal: JSON.stringify({ displayName, email, password }),
-        userEmail: email,
-      })
-    ) {
+    const SendSuccess = await SendVerification({
+      stringVal: JSON.stringify({ displayName, email, password }),
+      userEmail: email,
+    })
+    if (SendSuccess) {
       return NextResponse.json({ ok: true });
     } else {
       return NextResponse.json({ error: "Email send failed." }, { status: 502 });
